feat: add more trolling stuff

+++ src/types/script/ScriptEvent.cpp
@@ -5,6 +5,6 @@ void SCRIPT_EVENT::Send()
 {
 	auto size = __SizeOfEvent;
 	__SizeOfEvent = 0;
-	SCRIPT::_SEND_TU_SCRIPT_EVENT_NEW(1, this, size / sizeof(std::uint32_t), PlayerBits, EventIndex);
+	SCRIPT::_SEND_TU_SCRIPT_EVENT_NEW(1, this, size / sizeof(std::uint64_t), PlayerBits, EventIndex);
 	__SizeOfEvent = size;
 }
